fix outline when following
