+++ quadratic-client/src/app/ui/components/Following.tsx
@@ -1,41 +1,54 @@
 import { editorInteractionStateFollowAtom } from '@/app/atoms/editorInteractionStateAtom';
-import { MultiplayerUser } from '@/app/web-workers/multiplayerWebWorker/multiplayerTypes';
+import { useMultiplayerUsers } from '@/app/ui/menus/TopBar/useMultiplayerUsers';
+import { StopCircleIcon } from '@/shared/components/Icons';
 import { displayName } from '@/shared/utils/userUtil';
-import { StopCircleOutlined } from '@mui/icons-material';
-import { IconButton } from '@mui/material';
-import { useSetRecoilState } from 'recoil';
+import { useMemo } from 'react';
+import { useRecoilState } from 'recoil';
 
-interface FollowingProps {
-  follow?: MultiplayerUser;
-}
-
-export const Following = ({ follow }: FollowingProps) => {
-  const setFollow = useSetRecoilState(editorInteractionStateFollowAtom);
+export const Following = () => {
+  const [editorInteractionStateFollow, setFollow] = useRecoilState(editorInteractionStateFollowAtom);
+  const { users } = useMultiplayerUsers();
+  const follow = useMemo(
+    () =>
+      editorInteractionStateFollow ? users.find((user) => user.session_id === editorInteractionStateFollow) : undefined,
+    [editorInteractionStateFollow, users]
+  );
 
   if (!follow) return null;
 
   return (
-    <div
-      style={{
-        position: 'absolute',
-        top: 0,
-        left: '50%',
-        padding: 'calc(0.3rem - 3px) 1rem 0.3rem',
-        background: follow.colorString,
-        color: 'white',
-        fontSize: '0.85rem',
-        width: 'fit-content',
-        height: 'fit-content',
-        display: 'flex',
-        alignItems: 'center',
-        gap: '0.3rem',
-        transform: 'translateX(-50%)',
-      }}
-    >
-      <div>Following {displayName(follow, false)}</div>
-      <IconButton aria-label="stop" onClick={() => setFollow(undefined)} sx={{ padding: 0 }}>
-        <StopCircleOutlined sx={{ color: 'white' }} />
-      </IconButton>
-    </div>
+    <>
+      <div
+        className="flex items-center gap-3 rounded-r rounded-s px-4 py-1.5 text-sm text-white"
+        style={{
+          position: 'absolute',
+          top: 0,
+          left: '50%',
+          background: follow.colorString,
+          width: 'fit-content',
+          height: 'fit-content',
+          transform: 'translateX(-50%)',
+        }}
+      >
+        Following {displayName(follow, false)}
+        <button
+          aria-label="Stop following"
+          className="flex items-center opacity-80 hover:opacity-100"
+          onClick={() => setFollow(undefined)}
+        >
+          <StopCircleIcon />
+        </button>
+      </div>
+      <div
+        style={{
+          width: '100%',
+          height: '100%',
+          overflow: 'hidden',
+          position: 'absolute',
+          border: follow ? `3px solid ${follow.colorString}` : '',
+          pointerEvents: 'none',
+        }}
+      ></div>
+    </>
   );
 };
